updating tests based on functions

+++ tests/test_argon2.cc
@@ -12,7 +12,7 @@ int main() {
     std::cout << "passed: argon2id as " << result << "\n";
 
     auto verify = behh::cryptography_functions::hash::argon2id_verifier(password, result);
-    assert(verify == ARGON2_OK);
+    assert(verify == (int32_t)ARGON2_OK);
     std::cout << "passed: argon2id verified\n";
 #endif // BEHH_USING_ARGON2
